Fix: Dynamically import three.js libraries (#1166)
* dynamically import three js libraries

* move all models to models file

* smarter import

* specific import

* dynamic import of rigidbody

* better loading / color / comments
