+++ apps/web/app/(base-org)/(root)/page.tsx
@@ -17,16 +17,16 @@ import Link from 'apps/web/src/components/Link';
 import MissionSection from 'apps/web/src/components/base-org/root/MissionSection';
 import OpLogo from 'apps/web/public/images/op_logo.svg';
 
-const DynamicThreeHero = dynamic(async () => import('apps/web/src/components/ThreeHero'), {
-  ssr: false,
-});
-
+const Scene = dynamic(
+  async () => import('apps/web/src/components/ThreeHero').then((mod) => mod.Scene),
+  { ssr: false },
+);
 export default async function Home() {
   return (
     <ErrorsProvider context="base_landing_page">
       <AnalyticsProvider context="hero">
         <div className="relative z-10 h-screen w-full">
-          <DynamicThreeHero />
+          <Scene />
           <div className="absolute bottom-0 z-20 flex w-full flex-col justify-between gap-6 pb-20 text-white lg:flex-row">
             <div className="lg:ml-20">
               <Container>
